SIMD-117: create axios integration
