+++ src/app/masters/add/page.tsx
@@ -5,7 +5,7 @@ import FormComponent from '@/components/forms/FormComponent';
 import { Field } from 'formik';
 import { api } from '@/hooks/hooks-api';
 
-export default function AddDevicePage({openModal, setOpenModal}: {openModal: boolean, setOpenModal:Function}){
+export default function AddMasterPage({openModal, setOpenModal}: {openModal: boolean, setOpenModal:Function}){
 
     const [users, setUsers] = useState([]);
     const [masters, setMasters] = useState([]);
